Reference endpoints with filter options
